# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Desa Puundoho Information System.",
        "title": "Puundoho API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
        "/apbdes": {
            "get": {
                "description": "Get a list of all APBDes years",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "List APBDes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/main.ApbdDesa"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new APBDes entry (Bendahara only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "Create APBDes",
                "parameters": [
                    {
                        "description": "APBDes Data",
                        "name": "apbdes",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ApbdDesaInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.ApbdDesa"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/apbdes/pendapatan": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new pendapatan entry (Bendahara only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "Create Pendapatan",
                "parameters": [
                    {
                        "description": "Pendapatan Data",
                        "name": "pendapatan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPendapatanInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPendapatan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/apbdes/pendapatan/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing pendapatan entry by ID (Bendahara only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "Update Pendapatan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pendapatan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Data",
                        "name": "pendapatan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPendapatanInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPendapatan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a pendapatan entry by ID (Bendahara only)",
                "tags": [
                    "apbdes"
                ],
                "summary": "Delete Pendapatan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pendapatan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/apbdes/pengeluaran": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new pengeluaran entry (Bendahara only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "Create Pengeluaran",
                "parameters": [
                    {
                        "description": "Pengeluaran Data",
                        "name": "pengeluaran",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPengeluaranInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPengeluaran"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/apbdes/pengeluaran/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing pengeluaran entry by ID (Bendahara only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "Update Pengeluaran",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pengeluaran ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Data",
                        "name": "pengeluaran",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPengeluaranInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ApbdPengeluaran"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a pengeluaran entry by ID (Bendahara only)",
                "tags": [
                    "apbdes"
                ],
                "summary": "Delete Pengeluaran",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pengeluaran ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/apbdes/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an APBDes entry by ID (Bendahara only)",
                "tags": [
                    "apbdes"
                ],
                "summary": "Delete APBDes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "APBDes ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/apbdes/{id}/pendapatan": {
            "get": {
                "description": "Get a list of pendapatan for a specific APBDes ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "List Pendapatan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "APBDes ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/main.ApbdPendapatan"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/apbdes/{id}/pengeluaran": {
            "get": {
                "description": "Get a list of pengeluaran for a specific APBDes ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apbdes"
                ],
                "summary": "List Pengeluaran",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "APBDes ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/main.ApbdPengeluaran"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.ApbdDesa": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tahun": {
                    "type": "integer"
                },
                "total_pendapatan": {
                    "type": "integer"
                },
                "total_pengeluaran": {
                    "type": "integer"
                }
            }
        },
        "main.ApbdDesaInput": {
            "type": "object",
            "required": [
                "tahun"
            ],
            "properties": {
                "tahun": {
                    "type": "integer"
                },
                "total_pendapatan": {
                    "type": "integer"
                },
                "total_pengeluaran": {
                    "type": "integer"
                }
            }
        },
        "main.ApbdPendapatan": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_apbd": {
                    "type": "integer"
                },
                "jumlah": {
                    "type": "integer"
                },
                "kategori": {
                    "type": "string"
                }
            }
        },
        "main.ApbdPendapatanInput": {
            "type": "object",
            "required": [
                "id_apbd",
                "jumlah",
                "kategori"
            ],
            "properties": {
                "id_apbd": {
                    "type": "integer"
                },
                "jumlah": {
                    "type": "integer"
                },
                "kategori": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Paste your token with \"Bearer \" prefix (e.g. \"Bearer \u003ctoken\u003e\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}